refactor: split PopoverMainView to fix SwiftLint file_length + type_body_length
SwiftLint reported two violations on commit f126705:
  - file_length: 442 > 400 (warning)
  - type_body_length: PopoverMainView body 202 > 200 (warning)

Fix: extract header, local-runner row, action row, and inline job rows
into PopoverMainViewSubviews.swift. PopoverMainView.swift now contains
only the root struct (~115 lines, body ~40 lines). Subviews file is
~265 lines. No behaviour change; all 4 prior fixes remain intact.
